% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_d.R
\docType{data}
\name{cov_d}
\alias{cov_d}
\title{Covariance matrix for mlpca_d model}
\format{
A matrix with 20 rows and 20 columns
}
\usage{
cov_d
}
\description{
A random covariance matrix to simulate data errors
The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\references{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\keyword{datasets}
