% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\docType{data}
\name{data_clean}
\alias{data_clean}
\title{Error free data for all examples.}
\format{
A matrix with 300 rows and 20 columns
}
\source{
Wentzell, P. D., and S. Hou.
"Exploratory data analysis with noisy measurements."
Journal of Chemometrics 26.6 (2012): 264-281.
}
\usage{
data_clean
}
\description{
A dataset generated by the rotation of a bivariate normal density, the method
applied to get this dataset is described on Wentzell, P. D., and S. Hou.
"Exploratory data analysis with noisy measurements."
Journal of Chemometrics 26.6 (2012): 264-281.
}
\keyword{datasets}
