% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaned_mlpca_b.R
\docType{data}
\name{data_cleaned_mlpca_b}
\alias{data_cleaned_mlpca_b}
\title{Cleaned dataset after applied MLPCA B used for tests only}
\format{
A matrix with 300 rows and 20 columns
}
\source{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\usage{
data_cleaned_mlpca_b
}
\description{
A dataset where the values are estimated after mlpca_b is applied.
}
\keyword{datasets}
