% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaned_mlpca_c.R
\docType{data}
\name{data_cleaned_mlpca_c}
\alias{data_cleaned_mlpca_c}
\title{Cleaned dataset after applied MLPCA C used for tests only}
\format{
A matrix with 300 rows and 20 columns
}
\usage{
data_cleaned_mlpca_c
}
\description{
A dataset where the values are estimated after mlpca_c is applied.
}
\references{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\keyword{datasets}
