% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaned_mlpca_e.R
\docType{data}
\name{data_cleaned_mlpca_e}
\alias{data_cleaned_mlpca_e}
\title{Cleaned dataset after applied MLPCA E used for tests only}
\format{
A matrix with 30 rows and 20 columns
}
\usage{
data_cleaned_mlpca_e
}
\description{
A dataset where the values are estimated after mlpca_e is applied.
}
\references{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\keyword{datasets}
