% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_error_b.R
\docType{data}
\name{data_error_b}
\alias{data_error_b}
\title{Errors generated for mlpca_b model}
\format{
A matrix with 300 rows and 20 columns
}
\source{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\usage{
data_error_b
}
\description{
A dataset where each column contain values from a normal density with mean
= 0 and standard deviation from 0.2 to 1, the standard deviations differs in
the column. The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\keyword{datasets}
