% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_error_c.R
\docType{data}
\name{data_error_c}
\alias{data_error_c}
\title{Errors generated for mlpca_c model}
\format{
A matrix with 300 rows and 20 columns
}
\usage{
data_error_c
}
\description{
A dataset where each column contain values from a normal density with mean
= 0 and standard deviations simulated by a lognormal density with
meanlog = -4.75 and sdlog = 2.5, all the standard deviations are different.
The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\references{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\keyword{datasets}
