% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_error_e.R
\docType{data}
\name{data_error_e}
\alias{data_error_e}
\title{Errors generated for mlpca_e model}
\format{
A matrix with 30 rows and 20 columns
}
\usage{
data_error_e
}
\description{
A dataset where the values come from a 20 -multivariate normal density where
all the means are 0 and the covariance matrix from cov_e.
The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\references{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\keyword{datasets}
