% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpca_b.R
\name{mlpca_b}
\alias{mlpca_b}
\title{Maximum Likelihood Principal Component Analysis for Mode B Error
Conditions}
\usage{
mlpca_b(X, Xsd, p)
}
\arguments{
\item{X}{MxN matrix of measurements}

\item{Xsd}{MxN matrix of measurements error standard deviations}

\item{p}{Rank of the model's subspace, p must be than the minimum of M and N}
}
\value{
The parameters returned are the results of SVD on the estimated
subspace. The quantity Ssq represents the sum of squares of weighted
residuals.
}
\description{
Maximum Likelihood Principal Component Analysis for Mode B Error
Conditions
}
