% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpca_e.R
\encoding{UTF-8}
\name{mlpca_e}
\alias{mlpca_e}
\title{Maximum Likelihood Principal Component Analysis for Mode E Error
Conditions}
\usage{
mlpca_e(X, Cov, p)
}
\arguments{
\item{X}{IxJ matrix of measurements}

\item{Cov}{JXJXI matrices of measurement error covariance}

\item{p}{Rank of the model's subspace, p must be than the minimum of I and J}
}
\value{
The parameters returned are the results of SVD on the estimated
subspace. The quantity Ssq represents the sum of squares of weighted
residuals. ErrFlag indicates the convergence condition,
with 0 indicating normal termination and 1 indicating the maximum number of
iterations have been exceeded.
}
\description{
Performs maximum likelihood principal components analysis for
mode E error conditions (correlated errors, with a different covariance
matrix for each row, but no error correlation between the rows). Employs an
ALS algorithm.
}
\details{
The returned parameters, U, S and V, are analogs to the
truncated SVD solution, but have somewhat different properties since they
represent the MLPCA solution. In particular, the solutions for different
values of p are not necessarily nested (the rank 1 solution may not be in the
space of the rank 2 solution) and the eigenvectors do not necessarily account
for decreasing amounts of variance, since MLPCA is a subspace modeling
technique and not a variance modeling technique.
}
\references{
Wentzell, Peter D., et al. "Maximum likelihood principal
component analysis." Journal of Chemometrics: A Journal of the Chemometrics
Society 11.4 (1997): 339-366.
}
\author{
Renan Santos Barbosa
}
