% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_b.R
\docType{data}
\name{sds_b}
\alias{sds_b}
\title{Standard deviations for mlpca_b model}
\format{
A matrix with 300 rows and 20 columns
}
\source{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\usage{
sds_b
}
\description{
A dataset where each column contain the standard deviations from 0.2 to 1
that is necessary to run mlpca_b.
The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\keyword{datasets}
