% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_c.R
\docType{data}
\name{sds_c}
\alias{sds_c}
\title{Standard deviations for mlpca_c model}
\format{
A matrix with 300 rows and 20 columns
}
\source{
Wentzell, P. D. "Other topics in soft-modeling:
maximum likelihood-based soft-modeling methods." (2009): 507-558.
}
\usage{
sds_c
}
\description{
A dataset where each value come from a lognormal density with meanlog = -4.75
and sdlog = 2.5.
The main ideia is described in figure 3 on Wentzell, P. D.
"Other topics in soft-modeling: maximum likelihood-based soft-modeling
methods." (2009): 507-558.
}
\keyword{datasets}
