% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpca_b.R
\encoding{UTF-8}
\name{mlpca_b}
\alias{mlpca_b}
\title{Maximum Likelihood Principal Component Analysis for Mode B Error
Conditions}
\usage{
mlpca_b(X, Xsd, p)
}
\arguments{
\item{X}{MxN matrix of measurements}

\item{Xsd}{MxN matrix of measurements error standard deviations}

\item{p}{Rank of the model's subspace, p must be than the minimum of M and N}
}
\value{
The parameters returned are the results of SVD on the estimated
subspace. The quantity Ssq represents the sum of squares of weighted
residuals.
}
\description{
Performs maximum likelihood principal components analysis for
mode B error conditions (independent errors, homoscedastic within a column).
Equivalent to perfoming PCA on data scaled by the error SD, but results are
rescaled to the original space.
}
\details{
The returned parameters, U, S and V, are analogs to the
truncated SVD solution, but have somewhat different properties since they
represent the MLPCA solution. In particular, the solutions for different
values of p are not necessarily nested (the rank 1 solution may not be in the
space of the rank 2 solution) and the eigenvectors do not necessarily account
for decreasing amounts of variance, since MLPCA is a subspace modeling
technique and not a variance modeling technique.
}
\references{
Wentzell, Peter D., et al. "Maximum likelihood principal
component analysis." Journal of Chemometrics: A Journal of the Chemometrics
Society 11.4 (1997): 339-366.
}
\author{
Renan Santos Barbosa
}
